transaksi seeder

+++ database/migrations/2021_02_24_011441_create_keuangan_businesses_table.php
@@ -16,9 +16,9 @@ class CreateKeuanganBusinessesTable extends Migration
         Schema::create('keuangan_businesses', function (Blueprint $table) {
             $table->id();
             $table->foreignId('business_id')->constrained('businesses')->onDelete('cascade');
-            $table->double('pemasukan', 20, 2)->default(0);
-            $table->double('pengeluaran', 20, 2)->default(0);
-            $table->double('saldo', 20, 2)->default(10000000);
+            $table->decimal('pemasukan', 20, 2)->default(0);
+            $table->decimal('pengeluaran', 20, 2)->default(0);
+            $table->decimal('saldo', 30, 2)->default(10000000);
             $table->timestamps();
         });
     }
